RAG_Document_Intelligent/requirements.txt
numpy<2
openai
faiss-cpu
sentence-transformers==4.0.2
PyMuPDF
fastapi
uvicorn
requests
pipdeptree
python-multipart

torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
